+++ pylatex/figure.py
@@ -44,7 +44,7 @@ class Figure(Float):
         self.append(StandAloneGraphic(image_options=width,
                                       filename=fix_filename(filename)))
 
-    def _save_plot(self, *args, **kwargs):
+    def _save_plot(self, *args, extension='pdf', **kwargs):
         """Save the plot.
 
         Returns
@@ -55,13 +55,13 @@ class Figure(Float):
         import matplotlib.pyplot as plt
 
         tmp_path = make_temp_dir()
-        extension = kwargs.pop('extension', 'pdf').strip('.')
-        filename = posixpath.join(tmp_path, '{}.{}'.format(str(uuid.uuid4()), extension))
+        filename = '{}.{}'.format(str(uuid.uuid4()), extension.strip('.'))
+        filepath = posixpath.join(tmp_path, filename)
 
-        plt.savefig(filename, *args, **kwargs)
-        return filename
+        plt.savefig(filepath, *args, **kwargs)
+        return filepath
 
-    def add_plot(self, *args, **kwargs):
+    def add_plot(self, *args, extension='pdf', **kwargs):
         """Add the current Matplotlib plot to the figure.
 
         The plot that gets added is the one that would normally be shown when
@@ -71,6 +71,8 @@ class Figure(Float):
         ----
         args:
             Arguments passed to plt.savefig for displaying the plot.
+        extension : str
+            extension of image file indicating figure file type
         kwargs:
             Keyword arguments passed to plt.savefig for displaying the plot. In
             case these contain ``width`` or ``placement``, they will be used
@@ -84,7 +86,7 @@ class Figure(Float):
             if key in kwargs:
                 add_image_kwargs[key] = kwargs.pop(key)
 
-        filename = self._save_plot(*args, **kwargs)
+        filename = self._save_plot(*args, extension=extension, **kwargs)
 
         self.add_image(filename, **add_image_kwargs)